Apply mechanical clippy-driven cleanups
Ran `cargo clippy` across the crate and fixed the mechanical, low-risk
findings:
- or_insert_with(Vec::new) -> or_default() in build.rs and
  autotile/compiler.rs.
- Elided a redundant lifetime in autotile/matcher.rs::match_pattern.
- Rewrote a manual loop { match iter.next() { ... break } } as
  while let Some(...) = iter.next() in serialization.rs.
- Factored two "very complex type" flags into named type aliases
  (autotile/compiler.rs's CanonicalChecks, construction.rs's
  RestoreDesiredResult).
- Widened make_boring_room's &Vec<EorfInfo> param to &[EorfInfo].
- Dropped a redundant u8 cast in build_ui.rs and a needless .clone() on
  a Copy type in place.rs.
- headless.rs: args.get(0) -> args.first(), and a redundant &*deref.
- input.rs: added the missing #[allow(too_many_arguments)] to
  handle_furniture_pick (split out in the previous cleanup pass).
- qnn/translate.rs: simplified `v >= i + 1` to `v > i`.
- ui_util.rs: renamed FontColors::default() (shadowed the Default trait
  name) to FontColors::white(), and switched default_text_format() to
  struct-update syntax instead of a field assignment after
  Default::default().

Left the remaining clippy noise alone: `too_many_arguments` warnings are
either pre-existing Bevy systems or already carry an explicit allow;
`very complex type` on two Query/ParamSet types is normal Bevy style;
build_helpers.rs's dx/dy/dz casts turned out to be load-bearing for type
inference (removing them broke compilation), so they're back as they
were; two needless_range_loop suggestions were left alone because one
(qnn/translate.rs) relies on a documented embedding-length/constant
invariant that `.enumerate()` would silently stop enforcing, and the
other (build_ui.rs) is fiddly egui-closure code not worth the risk for
a style-only lint.

All 330 existing lib tests still pass.

+++ src/build_ui.rs
@@ -868,7 +868,7 @@ fn place_requirements_ui(
     visited.push(place_idx);
     for req in &places[place_idx].requirements {
         let count = match req.max {
-            Some(max) if max as u8 == req.min => format!("{}", req.min),
+            Some(max) if max == req.min => format!("{}", req.min),
             Some(max) => format!("{}-{}", req.min, max),
             None => format!("{}+", req.min),
         };
